requirement.txt
altair==4.1.0
appdirs==1.4.4
argon2-cffi==20.1.0
astor==0.8.1
attrs==20.1.0
backcall==0.2.0
base58==2.0.1
bleach==3.1.5
blinker==1.4
boto3==1.14.56
botocore==1.17.56
Brotli==1.0.9
cachetools==4.1.1
certifi==2020.6.20
cffi==1.14.2
chardet==3.0.4
click==7.1.2
colorama==0.4.3
dash==1.15.0
dash-core-components==1.11.0
dash-html-components==1.1.0
dash-renderer==1.7.0
dash-table==4.10.0
decorator==4.4.2
defusedxml==0.6.0
distlib==0.3.1
Django==2.2
docopt==0.6.2
docutils==0.15.2
entrypoints==0.3
enum-compat==0.0.3
filelock==3.0.12
Flask==1.1.2
Flask-Compress==1.5.0
future==0.18.2
idna==2.10
ipykernel==5.3.4
ipython==7.18.1
ipython-genutils==0.2.0
ipywidgets==7.5.1
itsdangerous==1.1.0
jedi==0.17.2
Jinja2==2.11.2
jmespath==0.10.0
joblib==0.16.0
jsonschema==3.2.0
jupyter-client==6.1.7
jupyter-core==4.6.3
MarkupSafe==1.1.1
mistune==0.8.4
nbconvert==5.6.1
nbformat==5.0.7
notebook==6.1.3
numpy==1.19.1
opencv-python==4.4.0.46
packaging==20.4
pandas==1.1.1
pandocfilters==1.4.2
parso==0.7.1
pathtools==0.1.2
pickleshare==0.7.5
Pillow==7.2.0
pipenv==2020.8.13
pipreqs==0.4.10
plotly==4.9.0
prometheus-client==0.8.0
prompt-toolkit==3.0.7
protobuf==3.13.0
pyarrow==1.0.1
pycparser==2.20
pydeck==0.5.0b1
Pygments==2.6.1
pyparsing==2.4.7
pyrsistent==0.16.0
pytesseract==0.3.7
python-dateutil==2.8.1
pytz==2020.1
pywin32==228
pywinpty==0.5.7
pyzmq==19.0.2
requests==2.24.0
retrying==1.3.3
s3transfer==0.3.3
scikit-learn==0.23.2
scipy==1.5.2
Send2Trash==1.5.0
six==1.15.0
sqlparse==0.3.1
streamlit==0.66.0
tabula==1.1.0
table-ocr==0.2.4
terminado==0.8.3
testpath==0.4.4
threadpoolctl==2.1.0
toml==0.10.1
toolz==0.10.0
tornado==6.0.4
traitlets==5.0.3
tzlocal==2.1
urllib3==1.25.10
validators==0.18.1
virtualenv==20.0.31
virtualenv-clone==0.5.4
virtualenvwrapper-win==1.2.6
watchdog==0.10.3
wcwidth==0.2.5
webencodings==0.5.1
Werkzeug==1.0.1
widgetsnbextension==3.5.1
yarg==0.1.9
